requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
bcrypt==4.1.2
dnspython==2.4.2